+++ src/EnumClass.Core/Infrastructure/Diagnostics.cs
@@ -1,6 +1,6 @@
 using Microsoft.CodeAnalysis;
 
-namespace EnumClass.Core;
+namespace EnumClass.Core.Infrastructure;
 
 public static class Diagnostics
 {
