+++ src/EnumClass.Generator/EnumClassIncrementalGenerator.cs
@@ -2,6 +2,8 @@
 using System.Diagnostics.CodeAnalysis;
 using System.Text;
 using EnumClass.Core;
+using EnumClass.Core.Infrastructure;
+using EnumClass.Core.Models;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
